improved tilde expansions, removed unnecessary functions
