+++ includes/ft_42sh.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_42sh.h                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: mrantil <mrantil@student.hive.fi>          +#+  +:+       +#+        */
+/*   By: mviinika <mviinika@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: Invalid date        by                   #+#    #+#             */
-/*   Updated: 2023/02/14 12:10:01 by mrantil          ###   ########.fr       */
+/*   Updated: 2023/02/14 15:11:42 by mviinika         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -511,9 +511,7 @@ char			*ft_expansion_dollar(t_shell *sh, char *str);
 char			*ft_expansion_tilde(t_shell *sh, char *str);
 char			*ft_expansion_excla(char *str, int i);
 void			ft_quote_blash_removal(char *buff);
-char			*user_expansions(char *str);
-char			*passwd_user(char *input);
-void			join_paths(char **user, char **temp, char **path, int opt);
+char			*user_expansions(char *input);
 
 /*				FT_TEST				*/
 int				ft_test_b(char **arg);
