+++ sources/builtins/ft_builtins.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_builtins.c                                      :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: mrantil <mrantil@student.hive.fi>          +#+  +:+       +#+        */
+/*   By: mviinika <mviinika@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: Invalid date        by                   #+#    #+#             */
-/*   Updated: 2023/02/13 13:25:18 by mrantil          ###   ########.fr       */
+/*   Updated: 2023/02/14 12:57:02 by mviinika         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
