+++ sources/parameter_formatting/find_from_end.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   find_from_end.c                                    :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: mbarutel <mbarutel@student.42.fr>          +#+  +:+       +#+        */
+/*   By: mviinika <mviinika@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/10 12:35:19 by mviinika          #+#    #+#             */
-/*   Updated: 2023/02/13 10:45:20 by mbarutel         ###   ########.fr       */
+/*   Updated: 2023/02/14 12:49:52 by mviinika         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,7 +18,6 @@ char	*find_from_end(char *haystack, char *needle)
 	int	len_needle;
 
 	len = (int)ft_strlen(haystack);
-	//ft_printf("haystack %s len [%d]]\n",haystack, len);
 	while (len > 0)
 	{
 		len_needle = (int)ft_strlen(needle);
@@ -32,8 +31,7 @@ char	*find_from_end(char *haystack, char *needle)
 			}
 			if (len_needle == 0)
 			{
-				//ft_printf(" haystack [%s] \n", &haystack[len]);
-				haystack = ft_strndup(haystack, ft_strlen(haystack) - ft_strlen(needle) + 1);
+				haystack = ft_strndup(haystack, ft_strlen(haystack) - ft_strlen(needle));
 				return (haystack);
 			}
 		}
