+++ sources/parameter_formatting/parameter_format.c
@@ -6,7 +6,7 @@
 /*   By: mviinika <mviinika@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/17 11:40:05 by mviinika          #+#    #+#             */
-/*   Updated: 2023/02/13 15:16:59 by mviinika         ###   ########.fr       */
+/*   Updated: 2023/02/14 13:02:15 by mviinika         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -71,6 +71,7 @@ static int	proceed_to_execute(t_param *pa, \
 	new_cmd = init_form_cycle(pa, cmd[ints->i], ints);
 	if (check_syntax(new_cmd))
 	{
+		ft_quote_blash_removal(new_cmd);
 		*ret = perform_param_expans(new_cmd, pa, &ints->ret);
 		substitute_og_cmd(pa, &cmd[ints->i], &ints->j);
 		if (*ret != 0)
